Introduce event handler levels
To make sure events are handled properly and reliably, the EEBUS core implementation should always be the first level of interacting with incoming SPINE messages, a middleware implementation second and the application last.

The EventHandler now provides these 3 levels and Subscribe and Unsubscribe expect the level to be provided. The Core level is only usable with an non public method for safeguard